+++ packages/adapters/src/index.ts
@@ -2,4 +2,3 @@ export * from "./mock";
 export * from "./factory";
 export * from "./base";
 export * from "./translate-gemma/local";
-export * from "./translate-gemma/huggingface";